backend/requirements.txt
# Smart Nutrition Platform - Requirements

# Core Framework
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.37
psycopg2-binary==2.9.10
asyncpg==0.31.0

# Data Validation
pydantic==2.10.6
pydantic-settings==2.7.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
bcrypt==4.2.1

# AI & Image Processing
google-generativeai==0.8.3
Pillow==11.1.0

# Dev Tools
pytest==8.3.5
pytest-asyncio==0.25.2
httpx==0.28.1
# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
